+++ vite.config.js
@@ -5,23 +5,12 @@ import react from '@vitejs/plugin-react';
 import dts from 'vite-plugin-dts';
 
 // https://vitejs.dev/config/
-export default defineConfig(({ command }) => {
-  if (command === 'serve') {
-    // Configuration for the development server
-    return {
-      plugins: [react()],
-      server: {
-        host: '0.0.0.0',
-        port: 5173,
-      },
-    };
-  } else {
+export default defineConfig(() => {
+  // Build the library if the VITE_BUILD_LIBRARY environment variable is set
+  if (process.env.VITE_BUILD_LIBRARY) {
     // Configuration for the library build
     return {
-      plugins: [
-        react(),
-        dts({ insertTypesEntry: true }),
-      ],
+      plugins: [react(), dts({ insertTypesEntry: true })],
       build: {
         lib: {
           entry: resolve(fileURLToPath(new URL('.', import.meta.url)), 'src/index.jsx'),
@@ -39,7 +28,26 @@ export default defineConfig(({ command }) => {
             exports: 'named',
           },
         },
+        // Prevent variable name mangling in GLSL shaders
+        terserOptions: {
+          compress: {
+            ecma: 2020,
+          },
+        },
       },
     };
   }
+
+  // Default configuration for the demo app (dev and preview)
+  return {
+    plugins: [react()],
+    // Configuration for the development server
+    server: {
+      host: '0.0.0.0',
+      port: 5173,
+    },
+    build: {
+      // Standard app build, which will produce an index.html
+    },
+  };
 });